pyproject.toml
[project]
name = "yelp-review-binary"
version = "0.1.0"
description = "Binary classification model for yelp-review dataset"
authors = [
    {name = "shpvckdotov",email = "korolev.do@phystech.edu"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "numpy (>=2.3.5,<3.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "dvc (>=3.64.2,<4.0.0)",
    "mlflow (>=3.7.0,<4.0.0)",
    "spacy (>=3.8.11,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "torch (==2.9.1)",
    "pytorch-lightning (>=2.6.0,<3.0.0)",
    "torchtext (>=0.18.0,<0.19.0)",
    "datasets (>=4.4.2,<5.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "transformers (>=4.57.3,<5.0.0)",
    "pyarrow (>=22.0.0,<23.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false 
[dependency-groups]
dev = [
    "fire (>=0.7.1,<0.8.0)",
    "pre-commit (>=4.5.1,<5.0.0)",
    "pylint (>=4.0.4,<5.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "dvc[s3] (>=3.64.2,<4.0.0)"
]
